2.16. Шаблонизируй это
destroy_after.

Отправка на 2 проверку
